Write function to find a signature in BIOS memory
As both ACPI and the MP spec (and probably other things) require
scanning BIOS memory regions to find specific signatures for their
data structures, a function has been created to perform this search.

This function is placed in a new bios.c file with a corresponding
header in asm/ as BIOS is architecture-specific.