Partial MP table parsing
Added functions to add locate and verify validity of the MP tables,
and implemented part of the parsing logic.
Rewrote the ACPI MADT structure functions to be generic, allowing them
to be shared between MADT and MP table parsing.
MP table parsing function is called in bsp_apic_init if MADT cannot be
found.