Rename rconfig sections
There is currently an issue in the rconfig cli frontend in which it uses
the current section name to detect whether the config section/file has
changed. This is a poor design choice and will be fixed later, but for
the time being the sections in the current rconfig have been renamed.

This renaming is useful regardless as these section names are now more
descriptive, instead of having a "General" section everywhere.