Add new console API
Created a new console API which allows the system to have multiple
virtual consoles. Each console is represented by a struct console which
stores its state and a struct consfn containing callback functions for
each of the console's actions.
All consoles in the system are stored in a list. One of these consoles
is marked as the active console. This console will be used by the
system for all text-based I/O.