Add move_cursor to console API
Added a move_cursor function to the console API which is used to change
the position of the cursor in the given console.
Implemented move_cursor for the VGA text console and wrote a function
which moves the visual VGA cursor on the screen. This function gets
called both from vgatext_move_cursor and from vgatxt_write to update the
position of the visual cursor after text is written to the VGA screen
buffer.