Parse buses and I/O APICs in a separate MP table walk
Instead of walking the table once to parse all entries, perform two
walks, the first reading bus and I/O APIC entries, and the second
reading processors and interrupts. The reason for this is that interrupt
entries rely on having bus and I/O APIC structures set up. If, for some
reason, a system has a table which places interrupts before I/O APICs or
buses, a one-pass parse would fail.