Do multiple MADT passes
Just like for the MP tables, added a function to walk the ACPI MADT and
call an entry handler on each entry. The MADT parsing code walks the
MADT twice, the first enumerating all I/O APICs in the system and
setting up their structures, and the second parsing every other entry.