Rewrite read_apic_id function
Since there is going to a be a struct lapic for each local APIC in the
system, the old read_apic_id function is no longer necessary. The
function has been renamed to find_cpu_lapic and it now reads the CPU's
APIC ID, finds the corresponding lapic structure and stores a pointer to
it in the local_apic per-CPU variable.