Exit read_cache_info if nreads == 0
There was a bug in the read_cache_info function where it assumed that
the first read of cpuid 2 would always return a number > 0 in the low
byte of eax (number of reads required). However, this is not a valid
assumption, as systems which do not support cpuid 2 could return a 0
to indicate that. In these situations, the code would get stuck in an
infinite loop. Another test for nreads == 0 has been added after the
initial cpuid instruction to check for this case. The return value of
the read_cache_info function has also been changed to a status code,
with this problematic scenario considered an error. The default cache
information is used when the function returns unsuccessfully.