DEFINE_PER_CPU_END assembler macro
The DEFINE_PER_CPU macro sets the assembler section to .percpu_data.
If the section is not reset after the variable is defined, all further
statements will end up in the .percpu_data section as well. To remove
the burden of resetting the section from the programmer, the
DEFINE_PER_CPU_END macro was added, which changes the section back to
.text.