Interrupt overhaul
Completely rewrote interrupt handling. IRQs (interrupt vectors) are now
resources which are allocated to devices that need them. A device will
request an interrupt using the new interrupt request API, and if all is
well, it will recieve an IRQ number and register a function to handle
the IRQ. To support older devices which are fixed to specific interrupt
pins, the interrupt API contains a request_fixed_irq function, which
allows the device to specify which pin it is connected to.
IRQ sharing has been introduced with this rewrite as well. Multiple
devices can share the same IRQ vector. When an IRQ is received, the
handler functions of all devices are which use that IRQ are called, and
each device will have to determine whether it sent the IRQ. The handling
of IRQ sharing is mandatory for fixed IRQ devices; for dynamic IRQs a
flag can be passed to the request_irq function to indicate that the
device allows IRQ sharing.