Define a basic initial timer framework
A timer is an abstraction of some object which maintains an incrementing
counter of "ticks" since its initialization. Reading a timer returns the
current value of the counter, and this value can be converted to a time
in nanoseconds through arithmetic operations. Timers will be used by the
radix kernel for low-level timekeeping purposes.

On boot, the kernel will determine which timers are available in the
system, and then, based on the timers' ratings, decide which should be
used. Ratings are static values hardcoded into each timer which
represent its overall usefulness.

Defined a struct timer which represents the timer abstraction, providing
the basic timer functions and attributes. Created a global variable to
store the active system timer, and added a function which registers a
new timer, updating the system timer as necessary.