ISR function fix
The ISR table is built using an array of function stubs which push
the interrupt vector then jump to commmon interrupt handling code,
with each entry 8 bytes long. This requires the use of the x86
push imm8 instruction, which is encoded in two bytes. The operand
of this instruction is a *signed* 8-bit value. The previous code
that generated the ISR stubs did not account for this and simply
pushed the vector number, resulting in a five byte push being used
after vector 0x7F, causing all subsequent entries to be 16 bytes long.

The number that is pushed has been corrected to always fit into a
signed 8-bit range by subtracting 0x80 from the vector number. It
is then immediately converted back to a vector number in the common
interrupt handling code.