Read number of local APIC LVT entries
The bug fixed in commit 42f36c1 was discovered due to a local APIC error
which occurred when the kernel was run on real hardware. Upon
investigation, the source of this error was found to be a write to the
CMCI LVT register, which does not exist in the target system's APIC,
but the APIC configuration code assumed that all 7 LVT entires exist.
The code has been modified to read the local APIC APICVER register to
obtain the number of LVT entries that the local APIC has, and then
only intialize those which exist.