Add time_ns function
Added the time_ns function, which returns the number of nanoseconds that
have passed since (shortly after) system boot. Initially, time_ns points
to a dummy function which always returns 0, but after the first system
timer is initialized, time_ns is updated to point to a function which
calculates elapsed nanoseconds based on the system timer's ticks.