Replace klog mutex with spinlock
The vklog function used a mutex to synchronize writes to the kernel log.
This was done because a mutex was the only synchronization primitive
that existed when it was written. Now that spinlocks exist, the klog
lock can be changed to one as it is more suitable for locking a short
and quick section of code, avoiding the overhead of a context switch.