Add mask_irq and unmask_irq
Added the mask_irq and unmask_irq functions to the irq.h API, which
expand to an architecture-specific method of handling IRQ masking.
The request_fixed_irq function no longer unmasks the IRQ by default
if it is successfully allocated; the device that requested the IRQ
has to unmask it manually.