Add ACPI power managment timer
Wrote a driver for the ACPI power management timer as a system timer
source. The driver maintains its own counter, and updates it with the
current value of the ACPI PM timer whenever read, accounting for timer
overflow. In order for the ACPI PM timer to function properly as a timer
source, it must be periodically read to prevent information loss. Since
the kernel does not have a system to manage events, this has temporarily
been set up using the PIT IRQ. Once a proper event system exists, the
PIT "event" code will be removed.