Start ACPI PM tick count when enabled
As the ACPI power management timer is always running, its tick count
runs from the point it was initialized when the system was first booted.
Changed this behaviour by reading the current tick count into the
previous tick count when the timer is enabled, so that it effectively
starts counting from that point.