Calculate timer mult/shift and max ticks
The frequency of some system timers is not known at compile time,
so mult and shift values must be calculated dynamically. A function
which performs this calculation has been added.

One important consideration to be made with timers is that, for high
values of mult, `ticks * mult` will eventually overflow a 64-bit
integer. The max_ticks and max_ns fields have been added to the struct
timer to indicate the point at which this overflow will occur. The new
timer_configure function calculates these values based on the timer's
mult and shift. The aforementioned function which cacluates mult and
shift also accounts for this by only choosing mult values which allow
a timer to run for 10 seconds without overflowing.

In the future, once the kernel gets a decent event system, an event will
be set up to accumulate the current system time based on the active
timer's tick count, and then reset the timer to prevent overflow.