Define new paging types for PAE
Created a new type paddr_t to represent physical addresses. This is
necessary as the size of a physical address on PAE systems is different
to that of a virtual address, and so the same time cannot be used to
reperesent the two (given that the virtual address type should always be
the same size as a pointer).

Defined the new pdpte_t type to represent entries in the page directory
pointer table in a PAE system. Updated the sizes of pteval_t and
pdeval_t for PAE.