Enable PAE on boot
When the kernel is compiled with the CONFIG_PAE option, set up all PAE
paging structures on boot and enable the PAE flag in CR4.
The _start assembly function has been changed to first jump to a
paging_init procedure, the definition of which is conditionally compiled
based on the CONFIG_PAE kernel configuration option. The old paging
setup code has been moved to the non-PAE version of this function.
The PAE version sets up the 3-level paging structures required for PAE,
recursively maps them, then enables PAE and paging before jumping to the
higher half kernel code.