Change phys_to_virt arg cast back to addr_t
The phys_to_virt function, by nature, can only work on physical
addresses in the kernel memory region, as they are at a constant
offset from their virtual addresses. Since virtual addresses on x86
are 32-bits, all physical addresses in the kernel memory region must
be small enough to not overflow when KERNEL_VIRTUAL_BASE is added to
them. Therefore, it is impossible for a kernel physical address to
ever be large enough to require (PAE-enabled) paddr_t to store it,
making a cast to paddr_t unnecessary.