Fix memory map parsing for mem > 4GiB
The init_region function took an addr_t as an argument, causing it to
fail with physical addresses greater than 0xFFFFFFFF. This has been
fixed by changing the type of the `base` argument to uint64_t.

A final zone_init call was being made after the next_phys_region loop,
with the intention of marking the final unused region of memory.
However, this function call used totalmem as an indication of the end of
memory, which does not account for physical memory gaps, such as the PCI
gap. With gaps, the value of `next` can be greater than totalmem,
causing an underflow which crashed the system. This zone_init call has
been removed, as it is unnecessary.