Fix bug in memory region length calculation
When aligning the base and length of a memory region, the length is
rounded down to PAGE_SIZE. To do this, it used the PAGE_MASK constant,
which is 32-bit on x86 systems. This resulted in the top half of the
64-bit length being cleared, causing the system to use much less memory
than available. The defintion of the PAGE_MASK constant has been inlined
to the length calculation, with an excplit cast to a 64-bit integer, to
make sure that it works correctly on all systems.