Correctly initialize user zone of physical memory
The user zone was being intialized starting at the end of the kernel
zone up to memsize pages. This was flawed as it did not account for
gaps in physical memory, such as the PCI gap, and could prevent large
regions of memory from entering the page allocator. To correct this, a
variable phys_mem_end was added, which indicates the last usable address
in phyiscal memory. The user zone is now initalized up to this address
instead of memsize.

When parsing memory regions, if a region expanded past MEM_LIMIT, it
would be ignored, potentially wasting available memory. This has been
changed so that the region is now truncated at MEM_LIMIT.