Add IRQ timers
An IRQ timer is an abstraction of a hardware timer which can produce
an IRQ at a terminal count. IRQ timers are a fundamental component of
radix, a tickless kernel, and will power radix's event subsystem.
Defined the irq_timer struct, describing an IRQ timer and its
operations, and wrote a function which sets the system's IRQ timer to
a specified source.