Added ISR for system events
Wrote a function which services event IRQs and calls a C event handler.
This is different to other ISRs as the values pushed onto the stack by
the interrupt may be modified in the C code if a context switch is
necessary. The ISR reuses code from the old interrupt handling routines
which accounts for ring 0 -> ring 0 interrupts in which esp and ss are
not pushed and reloaded.