Add reset function to system timers
Added a reset function to the timer struct and all x86 timers.
The reset function resets a timer's tick count to 0 and returns the
previous count. This is required to prevent overflow of shift * mult;
the system will periodically reset the system timer and accumulate
elapsed ticks.
In order to support tick resets, the functionalities of the HPET and
the ACPI PM timers have been rewritten. The ACPI PM has been reverted
to accumulate its own tick count, and the HPET now tracks ticks since
last reset, only returning the current difference when read.