Add timekeeping and dummy events
Wrote functions to manage the event queue and the scheduling of events.
Updated event_handler to schedule the next event in the queue after it
has finished running events. Created dummy events which are scheduled
when the next actual event would overflow the IRQ timer, and allocated
a dummy event struct per CPU. Added a timekeeping event which
accumulates timer ticks at a period of the system timer's max_ns / 2.