Event removal and timekeeping event update functions
Wrote a function which removes an event from the event queue, scheduling
a different timer IRQ if the removed event was at the front of the
queue. Added a function to update the period of the system timekeeping
event, to be used if the timer source needs to be changed.