Implement send_ipi function for the APIC
Wrote the lapic_send_ipi function, which uses the local APIC to send
an interprocessor interrupts to a set of processors. Added two wrapper
functions which manage sending an IPI to a group of processors
represented by a cpumask in both flat and cluster addressing modes,
and set the send_ipi member of the LAPIC's struct PIC depending on
addressing mode.