Refactor GDT code and add gdt_init_cpu
Extracted all of the common GDT initialization code to a separate
function called from each gdt_init_* function. Added a new function,
gdt_init_cpu, which creates and loads a GDT belonging to a specified
CPU number, to be used for GDT setup in SMP to avoid the chicken and
egg per-CPU variable problem. Added the base of the fs segment as an
argument to the gdt_init function.