Pass AP status to percpu_init
Added a parameter to the percpu_init and arch_percpu_init functions
which specifies whether the current CPU is an application processor
or not. This allows for per-CPU variables/structures to be initialized
only by APs if necessary (as the BSP's version is already set).
Wrote separate code paths in arch_percpu_init for the BSP and AP cases.