Refactored IRQ timer code to per-CPU support
Moved the schedule_timer_irq function out of event.c into timer.c
and changed it into a function pointer to one of two functions
depending on whether the current IRQ timer is per-CPU or not.
Added a similar pointer to a function which returns the max_ns
value of the system IRQ timer. Modified the event subsystem code
to use these new functions.