bits.h: add ffs, fix fls
Added the ffs function in both generic and i386-specific forms to
bits.h. Updated the existing fls function, making the following
changes: returned bits are 1-indexed, 0 is returned if bit is not
found, added support for 64-bit integers. Updated the log2 macro
to account for the new 1-indexed return of fls.