sched: add priority boosting task
Created a task which runs on each processor that boosts the priority of
other tasks in the processor's queues which have not run in a long time.
The priority boosting task runs in a loop, although this is not yet
perfect as the kernel has no sleeping functionality yet, causing it to
be scheduled as part of the normal scheduler cycle.

Fixed a nasty bug in the __prio_timeslice calculation function which
was using the incorrect multiplier for time periods.