panic: send IPI to halt other processors
Added a new panic IPI which makes a processor halt with IRQs disabled.
Updated the panic function to send this API to all other processors,
resulting in a complete system halt when panic is called.
Added a raw_write function to panic.c which forcefully bypasses the
active console's mutex to write to it directly. This is used from
within a panic context to ensure that the crash message is written.