panic: add stack traces
Added a stack trace to the output of a kernel panic. Currently, this
only prints addresses as the kernel does not have support for symbol
lookup. The stack trace is enabled with an rconfig setting.

Added debugging support to the Makefile, which is conditionally enabled
through the setting of the CONFIG_DEBUG option in rconfig. Compiling in
debug mode enables special debug-friendly compiler options.