Page allocator: fix zone size check
Moved the check to see if a zone fit a requested number of pages by
moving it up to the existing check to see if a zone is full, so that
a different zone can be chosen if it fails.
Fixed undefined behaviour in max_ord decrementing where it could
potentially underflow.