Global idle cpumask
This defines a cpumask of idle processors in the system and redefines
the is_idle function in terms of it. The idle task sets the CPU it is
running on as idle.

Additionally, a function to unset a CPU from the online cpumask is added
and used in the SMP boot code if a processor fails to initialize.