Generalize ISR handling and some interrupt code
Migrates common IRQs and the sched_wake IPI to use the same interrupt
context stack fixup that the event IRQ uses to have a consistent entry
point for interrupts. CPU exceptions do not yet use this; they will be
migrated in the future.

Additionally, creates a header file with interrupt vector definitions to
have a centralized place where all vectors can be viewed, instead of
having to look across several files. This file also explicitly defines
which vectors are reserved for use by the kernel.