VMM: Don't use NULL to represent the kernel space
This publicly exposes the kernel's address space and updates code to use
it instead of implicitly using NULL to represent it. kthreads now store
a valid VMM space pointer instead of NULL.

The kernel address space is also fully initialized with its architecture
paging base and context. This enables correctly switching back to the
kernel space if changing from a user task to a kthread.