sched: Remove SCHED_SELECT action
This changes task switches to only occur through the schedule()
function. It removes the old SELECT functionality which chose a task
without switching to it, relying on the interrupt handlers to perform
the context switch. This behavior was problematic as it required complex
handling in the ISR code, as well as having tasks with different saved
register states depending on how they switched, making switching back
difficult in some cases (e.g. a nested interrupt).

With this change, the schedule() function always performs a context
switch when a different task is selected. A new action, SCHED_PREEMPT,
is added, which only differs from SCHED_REPLACE in that it reconsiders
the current task as an option. (This was previously done via SELECT if
the task had time remaining.)

SCHED_SELECT was used in two parts of the code, which are updated:

  1. The sched_wake IPI handler is simply replaced with SCHED_PREEMPT.
  2. The event handler. The schedule call here cannot be replaced
     directly, as all events must be run before context switching.
     Instead, the handler tracks if a SCHED event is seen, and if so,
     calls schedule() once it has completed.

Register copying from interrupt handlers is removed as it is no longer
required. The ISR code is not yet updated; this will happen in a later
change.