Handle scheduling events with a past timestamp
It is possible for the kernel to miss a scheduled event handler
interrupt, then try to reschedule another event. If this happens, the
event it tries to schedule could already have passed its timestamp.

Previously, this possibility was ignored -- the event system assumed
that the timestamp was in the future. This led to underflow when
calculating the timer's tick count, resulting in the IRQ being scheduled
with a very long interval.

This fixes that issue by explicitly handling the case where a timestamp
has passed. The IRQ timer is scheduled to occur in one tick to run
immediately after interrupts are re-enabled.