Track if tasks are running on a CPU
This adds a bit to a task's flags field to indicate whether or not it is
running on a processor. Unlike the task's state, this bit is set at the
earliest moment the a picks up a task and cleared at the very last.

This could have several uses in the future. For now, it is used to avoid
the scheduler getting into a bad state when a task is unblocked
immediately after it blocks itself.