Don't try to use APIC timer if APIC is inactive
When testing running radix on my workstation, the APIC failed to
initialize (not sure why yet). This revealed another issue: the kernel
still tried to configure the APIC timer, resulting in a panic. The boot
code only checked whether the APIC was supported, not whether it was
active. This fixes that bug.